#39 refactor storage
